Move class header separation
